Implements Delete Blockchain data folder and resync operational, with additional security to only delete specific files and folers names

+++ modules/database.js
@@ -4,6 +4,7 @@ const datastore = require("nedb");
 const moment = require("moment");
 const path = require("path");
 const fs = require("fs");
+const fsextra = require("fs-extra");
 const os = require("os");
 
 let db;
@@ -185,36 +186,73 @@ ipcMain.on("deleteWalletData", (event, arg) => {
   });
 });
 
+/*
 ipcMain.on("deleteBlockchainData", (event, arg) => {
   var deleteFolderRecursive = function (path) {
     if (fs.existsSync(path)) {
       fs.readdirSync(path).forEach(function (file, index) {
         var curPath = path + "/" + file;
         if (fs.lstatSync(curPath).isDirectory()) {
+          // only deletes folders that have these names:
+          if (["geth", "keystores", "chaindata", "ethash", "lightchaindata", "nodes", "triecache"].includes(file)) {
           // recurse
           deleteFolderRecursive(curPath);
+          }
         } else {
-          // delete file
+          // delete file if name correspond to one of the geth files
+          if (fs.lstatSync(filePath).isFile() && (file.startsWith('LOCK') || file.startsWith('nodekey') || file.startsWith('transactions'))) {
           fs.unlinkSync(curPath);
+          }
         }
       });
       fs.rmdirSync(path);
     }
   };
 
-  function getBlockchainDataLocation() {
-    switch (os.type()) {
-      case "Darwin":
-        return path.join(os.homedir(), "Library", "Etica", "geth");
-        break;
-      default:
-        return path.join(process.env.APPDATA, "Etica", "geth");
-    }
-  }
-
   // delete folder in a synchronous recursive maner
-  deleteFolderRecursive(getBlockchainDataLocation());
+  deleteFolderRecursive(arg);
+
+  event.sender.send('initializeGethResponse', 2);
+
+
   event.returnValue = true;
+}); */
+
+ipcMain.on("deleteBlockchainData", async (event, arg) => {
+  try {
+    const deleteFolderRecursive = async function (dir) {
+      if (await fsextra.exists(dir)) {
+        const files = await fsextra.readdir(dir);
+
+        for (const file of files) {
+          const filePath = path.join(dir, file);
+          if ((await fsextra.lstat(filePath)).isDirectory()) {
+            if (["geth", "keystore", "chaindata", "ethash", "lightchaindata", "nodes", "triecache"].includes(file)) {
+              await fsextra.remove(filePath);
+              //console.log('removed', filePath);
+            }
+          } else {
+            if ( fsextra.lstat(filePath).isFile() && ["LOCK", "nodekey", "transactions"].some((prefix) => file.startsWith(prefix))) {
+              fsextra.unlinkSync(filePath);
+            }
+          }
+        }
+        
+        // dont remove main blockchain folder:
+        if(path.resolve(dir) != path.resolve(arg)){
+          fsextra.rmdirSync(dir);
+        }
+       
+      }
+    };
+
+    await deleteFolderRecursive(arg);
+  // console.log('Blockchain folder deleted with success');
+    event.sender.send("deleteBlockchainDataResponse", true);
+  } catch (error) {
+    console.error("Error deleting blockchain data:", error);
+    event.sender.send("deleteBlockchainDataResponse", false);
+  }
 });
